move test to all clients
